improve images navigation and a few tweaks in the modal style
