+++ src/hooks/utils/useImagesGallery.js
@@ -5,17 +5,21 @@ import { useCallback } from "react";
 function useImagesGallery(images) {
   const [handleOpenModal, handleCloseModal] = useModalContext();
 
-  console.log(images);
-
   const openImageFromIndex = useCallback(
     (idx) => {
-      console.log("open!", idx);
       const file = images[idx];
+
+      // go to next image or reset index to 0 if there are no subsequent images
+      const nextIndex = images.length > idx + 1 ? idx + 1 : 0;
+
+      // go to previous index or set to latest index if there are no preceding images
+      const prevIndex = idx === 0 ? images.length - 1 : idx - 1;
+
       handleOpenModal(ModalFullImage, {
         image: file,
         onClose: handleCloseModal,
-        onNext: () => openImageFromIndex(idx + 1),
-        onPrevious: () => openImageFromIndex(idx - 1),
+        onNext: () => openImageFromIndex(nextIndex),
+        onPrevious: () => openImageFromIndex(prevIndex),
         navigatorText: `${idx + 1}/${images.length}`
       });
     },
@@ -26,10 +30,3 @@ function useImagesGallery(images) {
 }
 
 export default useImagesGallery;
-
-// const openFullImageModal = (file) => {
-//     handleOpenModal(ModalFullImage, {
-//       image: file,
-//       onClose: handleCloseModal
-//     });
-//   };
